requirements.txt
# CTRP MCP Server Dependencies
# Genie MCP server on Azure Web Apps with Entra ID auth

# === Core Framework ===
fastapi>=0.115.0
uvicorn>=0.34.0
pydantic>=2.0.0

# === Databricks SDK ===
databricks-sdk>=0.65.0

# === Genie Integration ===
databricks-ai-bridge>=0.9.0

# === MCP Protocol (Streamable HTTP transport) ===
fastmcp>=2.14.0

# === Entra ID Token Validation ===
PyJWT[crypto]>=2.8.0
cryptography>=42.0.0
httpx>=0.27.0
